Move Notification queue to inmemory package
